rename lib test files
