+++ tests/unit/lib/page-info.test.js
@@ -1,7 +1,7 @@
 // @ts-check
 /* eslint-disable no-unused-vars */
-import { getButtonColorStringFromTabInfo } from '../../lib/index';
-import { q_module, q_test } from '../util';
+import { getButtonColorStringFromTabInfo } from '../../../lib/index';
+import { q_module, q_test } from '../../util';
 
 q_module('PageInfo utility tests', () => {
   q_test('getButtonColorStringFromTabInfo', async (assert) => {
